Optimization. Experimental offline calculation
Designed a simple method to catch the game up far more accurately. Offline progress does not account for non-linear relationships between resources.

For example, the relationship between sharks and fish is linear, and the relationship between nurse sharks and sharks is also linear. Therefore the relationship between nurse sharks and fish is quadratic; but the regular processor does not account for this, it simply raises the number of sharks and  then raises the number of fish according to those number of sharks.

This actually causes offline progress to be faster than normal progress, in this case. It's clear why this is an issue, and the way I tried solving it was by running all of the missing time at 20x the speed. But it's slow, so I'm disabling it. I'll need to find an alternate solution.

My current thoughts: Either I'll end up building an algorithm to determine the relationships between resources, or I'll ask around, because there are other devs who have encountered this problem.