Finished Implementing New Offline Progress System
Finished the new system. It turns out, it already accounted somewhat well for income caps. I've added a small loop at the start of the process to process the first 50 missing ticks completely, along with the last 50 ticks.

Everything in-between is handled by a sort of "dropping" algorithm. The h value is initially decided by how high timeDelta starts. Then the RK_4 function runs until a certain amount of time is left, returns that amount of time left, and it "drops" to the next algorithm in the list. It flows down the list until it reaches the end, where it processes the remaining 50 ticks at real speed, and processes any remainder between 0 and 1 linearly.

This algorithm is rather complex, but it is highly effective, and extremely accurate. I expect an error of only 0.5% - 3% when the time to process is between 12 and 48 hours. This has been verified experimentally using artificial exponential and natural quadratic relationships.

Glad to say, that means the offline progress system is done - and it's not the slowest thing in the world. 12-48 hours process in between 1 and 4 seconds on my own machine. that's a little while, so perhaps the value of h should be higher for that time frame, but I'm sure that nobody will mind by a significant amount waiting 6 or 7 or 10 seconds to play the game after a day or two of offline progress. It's still a minor problem though. I'll have to see how reactions to it are, then adjust accordingly.

Also, to Toby222, you're welcome to add yourself to the credits via a pull request. I'm not going to do it myself because I want you to decide where you want to add it, and what name you want to specify or links you want to add, etc.