Links
Added some links to the mod, added placeholder for changelog entry. Will probably be called New Frontiers, with no particular update name.

Not sure how to characterize in terms of the original update numbers. Should the mod just start at "0.8," or should it be "0.1" since it is its own thing? Should I say "0.71 modded?" It's a lot to think about...