New Sprite, Some Cleanup, Artifact Changes
Added sprite for eel sifter. Modified sponge biomimicry sprite again, am now satisfied with it.

Cleaned up a little bit of code here and there.

Disabled the terraformer *without* breaking save files in the process. It will no longer be available to purchase, has no effect, and cannot be seen in the reflection.

The only artifact change this time around will be to the distant foresight artifact, which now has a max level of 5, is more expensive, and has a more powerful effect. The foresight now reveals 20% of the features of worlds per level, rather than flat amounts of extra information. The distant foresight artifact will be important in the future, so its properties have been changed ahead of time.

The final pieces are coming together. Only a few more sprites, additional cleanup, and the changelog, which I will start now.